:recycle:Refact : UserManager를 새로운 managers.py 파일로 분리 #3
- 기존 models.py에 한꺼번에 작성한 코드를 분리하여 구조를 개선
