+++ src/restraunt_recommendation/users/models.py
@@ -1,21 +1,7 @@
-# Create your models here.
 from django.contrib.auth.models import AbstractBaseUser
-from django.contrib.auth.base_user import BaseUserManager
 from django.db import models
 
-
-class UserManager(BaseUserManager):
-    def create_user(self, username, password, **extra_fields):
-        if not username:
-            raise ValueError('계정명은 반드시 필요합니다!')
-        user = self.model(username=username, **extra_fields)
-        user.set_password(password)
-        user.save()
-        return user
-
-    def create_superuser(self, username, password, **extra_fields):
-        extra_fields.setdefault("is_admin", True)
-        return self.create_user(username, password, **extra_fields)
+from restraunt_recommendation.users.managers import UserManager
     
 
 class User(AbstractBaseUser):
